refactor(ipmi): extract IPMI tool into carbide-ipmi crate
Move the IPMI tool implementations out of the api crate into a new
standalone crate, carbide-ipmi, as part of the ongoing effort to
shrink the api mega-crate.

The new crate exposes only the IPMITool trait and three constructor
functions returning Arc<dyn IPMITool>:

  - carbide_ipmi::tool         (lanplus via /usr/bin/ipmitool)
  - carbide_ipmi::bmc_mock     (HTTP transport for bmc-mock)
  - carbide_ipmi::test_support (no-op for tests)

Concrete impls (IPMIToolImpl, IPMIToolHttpImpl, IPMIToolTestImpl) are
now private to their submodules; callers depend on the trait only.

Also cleans up IPMIToolImpl::new to take Option<u32> by value instead
of &Option<u32> (Option<u32> is Copy).

Callers updated: api (setup, site_explorer, state_controller,
test fixtures) and bmc-explorer-cli.

No behavior change.
